refactor(restaurant): store 앱 이름 변경
- store -> restaurant
