+++ apps/restaurant/admin.py
